Add automatic CCL list generation for prefill and decode when user does not provide lists (#663)
For using CCL feature during prefilling and decoding, the user needs to
pass two lists of context lengths to be used during these processes.
Here, we add an option to generate these lists automatically when they
are not provided by the user manually. These list generation is suitable
for a general-purpose application which considers both prefilling and
decoding processes and generates CCL lists for both.

---------

+++ QEfficient/generation/text_generation_inference.py
@@ -819,7 +819,9 @@ class QEffTextGenerationBase:
                 inputs["lora_ids"] = np.array(batch_lora_ids, dtype=np.int64).reshape(self.batch_size, 1)
 
         if self.comp_ctx_lengths_prefill is not None:
-            self.list_of_comp_ctx_lengths_prefill = [np.zeros(length) for length in self.comp_ctx_lengths_prefill]
+            self.list_of_comp_ctx_lengths_prefill = [
+                np.zeros(length, dtype=np.int8) for length in self.comp_ctx_lengths_prefill
+            ]
             prefill_ccl_id = 0
             inputs["comp_ctx_lengths"] = self.list_of_comp_ctx_lengths_prefill[prefill_ccl_id]
 
@@ -849,7 +851,9 @@ class QEffTextGenerationBase:
         )
 
     def initialize_ccl(self, decode_inputs):
-        self.list_of_comp_ctx_lengths_decode = [np.zeros(length) for length in self.comp_ctx_lengths_decode]
+        self.list_of_comp_ctx_lengths_decode = [
+            np.zeros(length, dtype=np.int8) for length in self.comp_ctx_lengths_decode
+        ]
         max_ccl_id = len(self.comp_ctx_lengths_decode) - 1
         max_position_id = np.max(decode_inputs["position_ids"])
         ccl_id_initial = 0
